Add better parsing logic for NetworkManager passkey
1. Terminate parsing when we see a closing char ' or " that matches the opening char

There was an issue of prematurely cutting off a quoted value.
Ex. "some'one is watching" -> "some"
Now: When parsing a quoted value (" or '), we stop parsing when we see a corresponding unescaped quote.

2. Add additional \ for an \ that appear in passkey for NetworkManager

NetworkManager specifically needs \ to be escaped. Other special characters seem to be fine.

3. No longer strip whitespace from return val

Ex. If my psk is "  password with spaces to left and right   " in piaware-config.txt. We need to retain the spaces in the quotes.
Whitespaces not enclosed in quotes will still be removed.

4. Any un-escaped ", ' or \ may lead to issues

Ex. "pass'word" is fine. "pass"word" is obviously not.